cosmetics

+++ run.ps1
@@ -27,11 +27,14 @@ if (-not $PythonExe) {
 
 Write-Host "[+] Python found: $($PythonExe.Source)" -ForegroundColor Green
 
+$is_venv = $false
+
 # 3. Virtual Environment management
 if (-not (Test-Path $VenvPath)) {
     Write-Host "[2/4] Creating virtual environment..." -ForegroundColor Cyan
     & $PythonExe.Source -m venv $VenvPath
 } else {
+    $is_venv = $true
     Write-Host "[2/4] Virtual environment already exists." -ForegroundColor Green
 }
 
@@ -52,7 +55,7 @@ if ($env:VIRTUAL_ENV -and (Test-Path $env:VIRTUAL_ENV)) {
 
 # 5. Execution
 try {
-    if (Test-Path $Requirements) {
+    if ((Test-Path $Requirements) -and -not $is_venv) {
         Write-Host "[+] Installing requirements..." -ForegroundColor Yellow
         python -m pip install --upgrade pip --quiet
         pip install -r $Requirements
@@ -62,7 +65,7 @@ try {
     }
     Write-Host "[4/4] Starting $AppName..." -ForegroundColor Cyan
     Write-Host "-----------------------------------------------------------" -ForegroundColor Gray
-    python $PythonScript
+    python $PythonScript $args
 }
 catch {
     Write-Host "An error occurred during execution." -ForegroundColor Red
